requirements.txt
appnope==0.1.2
asgiref==3.3.1
attrs==20.3.0
backcall==0.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
decorator==5.0.7
defusedxml==0.7.1
Django==3.1.7
django-allauth==0.44.0
django-anymail==8.2
django-autoslug==1.9.8
django-crispy-forms==1.11.1
django-environ==0.4.5
django-extensions==3.1.2
django-shopping-cart==0.1
idna==2.10
iniconfig==1.1.1
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
jsonfield==3.1.0
oauthlib==3.1.0
packaging==20.9
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
pluggy==0.13.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
py==1.10.0
pycparser==2.20
pycryptodome==3.10.1
Pygments==2.8.1
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.3
pytest-django==4.2.0
python3-openid==3.2.0
pytz==2021.1
rave-python==1.2.10
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
stripe==2.56.0
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
